requirements.txt
gunicorn==19.7.1
requests==2.18.4
Flask==0.12.2
lxml==3.8.0
python-dateutil==2.6.1
nltk==3.2.4
beautifulsoup4==4.6.0
